+++ flutter/lib/widgets/tab_bar_widget.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:ru_project/services/api_service.dart';
 import 'package:ru_project/services/logger.dart';
 import 'package:provider/provider.dart';
+import 'package:ru_project/widgets/old_map_widget.dart';
 import 'package:ru_project/widgets/map_widget.dart';
-import 'package:ru_project/widgets/ru_lumiere_sector_map.dart';
 import 'package:ru_project/widgets/menu_widget.dart';
 import 'package:ru_project/models/color.dart';
 import 'package:ru_project/providers/user_provider.dart';
